[FEAT] Pose Estimation VC 이식 (#27)
